TSP-612-Debugger never hits breakpoint (#30)
* breakpoint lineNumber changed to u32

* updating changelog

+++ src/debugger/breakpoint/mod.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct Breakpoint {
     #[serde(rename = "LineNumber")]
-    pub line_number: u8,
+    pub line_number: u32,
     #[serde(rename = "Enable")]
     pub enable: bool,
     #[serde(rename = "Condition")]
